Move MCP upstream primitives into plugins service
Interfaces:

- services/plugins/mcpupstream.New(ctx, name, url, mode, headers, checkEgress, opts...) constructs MCP-backed providers without internal imports.

- (*mcpupstream.Upstream).FilterOperations(map[string]*operationexposure.OperationOverride) applies allowed_operations policy to upstream catalogs and tool calls.

- operationexposure.New(allowed) returns a Policy with ValidateCatalog, ApplyCatalog, Resolve, and Wrap helpers.

- operationexposure.NewRestricted preserves existing restricted-provider behavior, with services/plugins/declarative.NewRestricted left as a compatibility facade.

- services/plugins/headerutil.NormalizeHeaders and MergeHeaders own plugin header canonicalization and request-token precedence.

Example:

up, err := mcpupstream.New(ctx, "notion", url, core.ConnectionModeUser, headers, checkEgress)

err = up.FilterOperations(map[string]*operationexposure.OperationOverride{"search": {Alias: "notion.search"}})

policy, err := operationexposure.New(allowedOperations); provider = policy.Wrap(provider)
