+++ gestaltd/services/invocation/direct_tool.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/valon-technologies/gestalt/server/core"
-	"github.com/valon-technologies/gestalt/server/internal/mcpupstream"
 	"github.com/valon-technologies/gestalt/server/services/identity/principal"
+	"github.com/valon-technologies/gestalt/server/services/plugins/mcpupstream"
 
 	mcpgo "github.com/mark3labs/mcp-go/mcp"
 )
